+++ _since_time_text.d.ts
@@ -1,0 +1,8 @@
+/**
+ *
+ * @param date
+ * @returns {string}
+ * @see {@link https://stackoverflow.com/questions/3177836/how-to-format-time-since-xxx-e-g-4-minutes-ago-similar-to-stack-exchange-site#answer-3177838}
+ */
+export declare function _since_time_text(date: any): string;
+export { _since_time_text as _text__time__since };