+++ _timedout.d.ts
@@ -1,0 +1,1 @@
+export declare function _timedout(start: Date, timout_ms: number): boolean;