+++ _timedout.ts
@@ -1,0 +1,3 @@
+export function _timedout(start: Date, timout_ms: number) {
+	return new Date().getTime() >= (start.getTime() + timout_ms)
+}