+++ _timestamp_milliseconds.d.ts
@@ -1,0 +1,2 @@
+export declare function _timestamp_milliseconds(): number;
+export { _timestamp_milliseconds as _milliseconds__timestamp };