+++ day_milliseconds.d.ts
@@ -1,0 +1,2 @@
+export declare const day_milliseconds: number;
+export { day_milliseconds as milliseconds__day };