+++ day_milliseconds.ts
@@ -1,0 +1,5 @@
+import { day_seconds } from './day_seconds'
+export const day_milliseconds = day_seconds * 1000
+export {
+	day_milliseconds as milliseconds__day
+}