+++ hour_milliseconds.d.ts
@@ -1,0 +1,2 @@
+export declare const hour_milliseconds: number;
+export { hour_milliseconds as milliseconds__hour };