+++ hour_milliseconds.ts
@@ -1,0 +1,5 @@
+import { hour_seconds } from './hour_seconds'
+export const hour_milliseconds = hour_seconds * 1000
+export {
+	hour_milliseconds as milliseconds__hour
+}