+++ minute_milliseconds.d.ts
@@ -1,0 +1,2 @@
+export declare const minute_milliseconds: number;
+export { minute_milliseconds as milliseconds__minute };