+++ minute_milliseconds.ts
@@ -1,0 +1,5 @@
+import { minute_seconds } from './minute_seconds'
+export const minute_milliseconds = minute_seconds * 1000
+export {
+	minute_milliseconds as milliseconds__minute
+}