+++ minute_seconds.ts
@@ -1,0 +1,4 @@
+export const minute_seconds = 60
+export {
+	minute_seconds as seconds__minute
+}