+++ month_milliseconds.d.ts
@@ -1,0 +1,2 @@
+export declare const month_milliseconds: number;
+export { month_milliseconds as milliseconds__month };