+++ month_milliseconds.ts
@@ -1,0 +1,5 @@
+import { month_seconds } from './month_seconds'
+export const month_milliseconds = month_seconds * 1000
+export {
+	month_milliseconds as milliseconds__month
+}