+++ year_milliseconds.d.ts
@@ -1,0 +1,2 @@
+export declare const year_milliseconds: number;
+export { year_milliseconds as milliseconds__year };