+++ year_milliseconds.ts
@@ -1,0 +1,5 @@
+import { year_seconds } from './year_seconds'
+export const year_milliseconds = year_seconds * 1000
+export {
+	year_milliseconds as milliseconds__year
+}